+++ climpred/bootstrap.py
@@ -51,7 +51,66 @@ def _resample(hind, resample_dim):
     return smp_hind
 
 
-def _distribution_to_ci(ds, ci_low, ci_high, dim='bootstrap'):
+def _resample_iterations_idx(init, iterations, dim='member', replace=True):
+    """Resample over ``dim`` by index ``iterations`` times.
+
+    .. note::
+        This is a much faster way to bootstrap than resampling each iteration
+        individually and applying the function to it. However, this will create a
+        DataArray with dimension ``iteration`` of size ``iterations``. It is probably
+        best to do this out-of-memory with ``dask`` if you are doing a large number
+        of iterations or using spatial output (i.e., not time series data).
+
+    Args:
+        init (xr.DataArray, xr.Dataset): Initialized prediction ensemble.
+        iterations (int): Number of bootstrapping iterations.
+        dim (str): Dimension name to bootstrap over. Defaults to ``'member'``.
+        replace (bool): Bootstrapping with or without replacement. Defaults to ``True``.
+
+    Returns:
+        xr.DataArray, xr.Dataset: Bootstrapped data with additional dim ```iteration```
+
+    """
+
+    def select_bootstrap_indices_ufunc(x, idx):
+        """Selects multi-level indices ``idx`` from xarray object ``x`` for all
+        iterations."""
+        # `apply_ufunc` sometimes adds a singleton dimension on the end, so we squeeze
+        # it out here. This leverages multi-level indexing from numpy, so we can
+        # select a different set of, e.g., ensemble members for each iteration and
+        # construct one large DataArray with ``iterations`` as a dimension.
+        return np.moveaxis(x.squeeze()[idx.squeeze().transpose()], 0, -1)
+
+    # resample with or without replacement
+    if replace:
+        idx = np.random.randint(0, init[dim].size, (iterations, init[dim].size))
+    elif not replace:
+        # create 2d np.arange()
+        idx = np.linspace(
+            (np.arange(init[dim].size)),
+            (np.arange(init[dim].size)),
+            iterations,
+            dtype='int',
+        )
+        # shuffle each line
+        for ndx in np.arange(iterations):
+            np.random.shuffle(idx[ndx])
+    idx_da = xr.DataArray(
+        idx,
+        dims=('iteration', dim),
+        coords=({'iteration': range(iterations), dim: init[dim]}),
+    )
+
+    return xr.apply_ufunc(
+        select_bootstrap_indices_ufunc,
+        init.transpose(dim, ...),
+        idx_da,
+        dask='parallelized',
+        output_dtypes=[float],
+    )
+
+
+def _distribution_to_ci(ds, ci_low, ci_high, dim='iteration'):
     """Get confidence intervals from bootstrapped distribution.
 
     Needed for bootstrapping confidence intervals and p_values of a metric.
@@ -60,7 +119,7 @@ def _distribution_to_ci(ds, ci_low, ci_high, dim='bootstrap'):
         ds (xarray object): distribution.
         ci_low (float): low confidence interval.
         ci_high (float): high confidence interval.
-        dim (str): dimension to apply xr.quantile to. Default: 'bootstrap'
+        dim (str): dimension to apply xr.quantile to. Default: 'iteration'
 
     Returns:
         uninit_hind (xarray object): uninitialize hindcast with hind.coords.
@@ -88,7 +147,7 @@ def _pvalue_from_distributions(simple_fct, init, metric=None):
         pv (xarray object): probability that simple forecast performs better
                             than initialized forecast.
     """
-    pv = ((simple_fct - init) > 0).sum('bootstrap') / init.bootstrap.size
+    pv = ((simple_fct - init) > 0).sum('iteration') / init.iteration.size
     if not metric.positive:
         pv = 1 - pv
     return pv
@@ -259,9 +318,9 @@ def _bootstrap_by_stacking(init_pm, control):
 
 
 def _bootstrap_func(
-    func, ds, resample_dim, sig=95, bootstrap=500, *func_args, **func_kwargs,
+    func, ds, resample_dim, sig=95, iterations=500, *func_args, **func_kwargs,
 ):
-    """Sig percent threshold of function based on resampling with replacement.
+    """Sig % threshold of function based on iterations resampling with replacement.
 
     Reference:
     * Mason, S. J., and G. M. Mimmack. “The Use of Bootstrap Confidence
@@ -276,7 +335,7 @@ def _bootstrap_func(
             CPUs available.
         resample_dim (str): dimension to resample from.
         sig (int,float,list): significance levels to return. Defaults to 95.
-        bootstrap (int): number of resample iterations. Defaults to 500.
+        iterations (int): number of resample iterations. Defaults to 500.
         *func_args (type): `*func_args`.
         **func_kwargs (type): `**func_kwargs`.
 
@@ -292,19 +351,18 @@ def _bootstrap_func(
     else:
         psig = sig / 100
 
-    bootstraped_results = []
-    for _ in range(bootstrap):
-        smp_init_pm = _resample(ds, resample_dim)
-        bootstraped_results.append(func(smp_init_pm, *func_args, **func_kwargs))
-    sig_level = xr.concat(bootstraped_results, dim='bootstrap', **CONCAT_KWARGS)
-    sig_level = rechunk_to_single_chunk_if_more_than_one_chunk_along_dim(
-        sig_level, dim='bootstrap'
+    bootstraped_ds = _resample_iterations_idx(
+        ds, iterations, dim=resample_dim, replace=False
     )
-    sig_level = sig_level.quantile(dim='bootstrap', q=psig, skipna=False)
+    bootstraped_results = func(bootstraped_ds, *func_args, **func_kwargs)
+    bootstraped_results = rechunk_to_single_chunk_if_more_than_one_chunk_along_dim(
+        bootstraped_results, dim='iteration'
+    )
+    sig_level = bootstraped_results.quantile(dim='iteration', q=psig, skipna=False)
     return sig_level
 
 
-def dpp_threshold(control, sig=95, bootstrap=500, dim='time', **dpp_kwargs):
+def dpp_threshold(control, sig=95, iterations=500, dim='time', **dpp_kwargs):
     """Calc DPP significance levels from re-sampled dataset.
 
     Reference:
@@ -318,11 +376,11 @@ def dpp_threshold(control, sig=95, bootstrap=500, dim='time', **dpp_kwargs):
         * climpred.stats.dpp
     """
     return _bootstrap_func(
-        dpp, control, dim, sig=sig, bootstrap=bootstrap, **dpp_kwargs
+        dpp, control, dim, sig=sig, iterations=iterations, **dpp_kwargs
     )
 
 
-def varweighted_mean_period_threshold(control, sig=95, bootstrap=500, time_dim='time'):
+def varweighted_mean_period_threshold(control, sig=95, iterations=500, time_dim='time'):
     """Calc the variance-weighted mean period significance levels from re-sampled
     dataset.
 
@@ -331,7 +389,7 @@ def varweighted_mean_period_threshold(control, sig=95, bootstrap=500, time_dim='
         * climpred.stats.varweighted_mean_period
     """
     return _bootstrap_func(
-        varweighted_mean_period, control, time_dim, sig=sig, bootstrap=bootstrap,
+        varweighted_mean_period, control, time_dim, sig=sig, iterations=iterations,
     )
 
 
@@ -345,7 +403,7 @@ def bootstrap_compute(
     dim='init',
     resample_dim='member',
     sig=95,
-    bootstrap=500,
+    iterations=500,
     pers_sig=None,
     compute=compute_hindcast,
     resample_uninit=bootstrap_uninitialized_ensemble,
@@ -370,7 +428,7 @@ def bootstrap_compute(
                    initialized skill. Defaults to 95.
         pers_sig (int): Significance level for persistence skill confidence levels.
                         Defaults to sig.
-        bootstrap (int): number of resampling iterations (bootstrap
+        iterations (int): number of resampling iterations (bootstrap
                          with replacement). Defaults to 500.
         compute (func): function to compute skill.
                         Choose from
@@ -443,7 +501,7 @@ def bootstrap_compute(
     isHindcast = True if comparison in HINDCAST_COMPARISONS else False
     reference_alignment = alignment if isHindcast else 'same_inits'
 
-    for i in range(bootstrap):
+    for i in range(iterations):
         # resample with replacement
         smp_hind = _resample(hind, resample_dim)
         # compute init skill
@@ -495,11 +553,11 @@ def bootstrap_compute(
                     **metric_kwargs,
                 )
             )
-    init = xr.concat(init, dim='bootstrap', **CONCAT_KWARGS)
-    uninit = xr.concat(uninit, dim='bootstrap', **CONCAT_KWARGS)
+    init = xr.concat(init, dim='iteration', **CONCAT_KWARGS)
+    uninit = xr.concat(uninit, dim='iteration', **CONCAT_KWARGS)
     # when persistence is not computed set flag
     if pers != []:
-        pers = xr.concat(pers, dim='bootstrap', **CONCAT_KWARGS)
+        pers = xr.concat(pers, dim='iteration', **CONCAT_KWARGS)
         pers_output = True
     else:
         pers_output = False
@@ -535,7 +593,7 @@ def bootstrap_compute(
     if 'init' in init_skill:
         init_skill = init_skill.mean('init')
     # uninit skill as mean resampled uninit skill
-    uninit_skill = uninit.mean('bootstrap')
+    uninit_skill = uninit.mean('iteration')
     if not metric.probabilistic:
         pers_skill = reference_compute(
             hind, verif, metric=metric, alignment=reference_alignment, **metric_kwargs
@@ -574,7 +632,7 @@ def bootstrap_compute(
     # Attach climpred compute information to skill
     metadata_dict = {
         'confidence_interval_levels': f'{ci_high}-{ci_low}',
-        'bootstrap_iterations': bootstrap,
+        'bootstrap_iterations': iterations,
         'p': 'probability that uninitialized ensemble performs better than initialized',
         'reference_compute': reference_compute.__name__,
     }
@@ -607,7 +665,7 @@ def bootstrap_hindcast(
     dim='init',
     resample_dim='member',
     sig=95,
-    bootstrap=500,
+    iterations=500,
     pers_sig=None,
     reference_compute=compute_persistence,
     **metric_kwargs,
@@ -631,7 +689,7 @@ def bootstrap_hindcast(
                    initialized skill. Defaults to 95.
         pers_sig (int): Significance level for persistence skill confidence levels.
                         Defaults to sig.
-        bootstrap (int): number of resampling iterations (bootstrap
+        iterations (int): number of resampling iterations (bootstrap
                          with replacement). Defaults to 500.
         reference_compute (func): function to compute a reference forecast skill with.
                         Default: :py:func:`climpred.prediction.compute_persistence`.
@@ -713,7 +771,7 @@ def bootstrap_hindcast(
         dim=dim,
         resample_dim=resample_dim,
         sig=sig,
-        bootstrap=bootstrap,
+        iterations=iterations,
         pers_sig=pers_sig,
         compute=compute_hindcast,
         resample_uninit=bootstrap_uninitialized_ensemble,
@@ -730,7 +788,7 @@ def bootstrap_perfect_model(
     dim=None,
     resample_dim='member',
     sig=95,
-    bootstrap=500,
+    iterations=500,
     pers_sig=None,
     reference_compute=compute_persistence,
     **metric_kwargs,
@@ -754,7 +812,7 @@ def bootstrap_perfect_model(
                    initialized skill. Defaults to 95.
         pers_sig (int): Significance level for persistence skill confidence levels.
                         Defaults to sig.
-        bootstrap (int): number of resampling iterations (bootstrap
+        iterations (int): number of resampling iterations (bootstrap
                          with replacement). Defaults to 500.
         reference_compute (func): function to compute a reference forecast skill with.
                         Default: :py:func:`climpred.prediction.compute_persistence`.
@@ -819,7 +877,7 @@ def bootstrap_perfect_model(
         dim=dim,
         resample_dim=resample_dim,
         sig=sig,
-        bootstrap=bootstrap,
+        iterations=iterations,
         pers_sig=pers_sig,
         compute=compute_perfect_model,
         resample_uninit=bootstrap_uninit_pm_ensemble_from_control_cftime,