+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ What's New
 climpred v2.1.0 (2020-04-##)
 ============================
 
+Breaking change
+---------------
+
+- replace keyword `bootstrap` with `iterations` (:pr:`354`) `Aaron Spring`_.
+
+
 New Features
 ------------
 
@@ -67,6 +73,9 @@ Internals/Minor Fixes
   (:pr:`351`) `Aaron Spring`_.
 - Require ``xskillscore v0.0.15`` and use their functions for effective sample
   size-based metrics. (:pr: `353`) `Riley X. Brady`_.
+- Faster bootstrapping without replacement used in threshold functions of
+  climpred.stats (:pr:`354`) `Aaron Spring`_.
+
 
 
 Documentation